pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "HDRP"
version = "0.1.0"
description = "High-fidelity Digitized Research Platform"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "protobuf>=4.21.0",
    "openai>=1.0.0",
    "sentry-sdk>=1.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]
onnx = [
    "onnx>=1.14.0",
    "onnxruntime>=1.16.0",
]
onnx-gpu = [
    "onnx>=1.14.0",
    "onnxruntime-gpu>=1.16.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["HDRP*"]
